# Optimize `MCP` by removing `assoc` usage