--- a/src/analyses/mCP.ml
+++ b/src/analyses/mCP.ml
@@ -161,7 +161,7 @@ struct
         f ((k,v::a')::a) b
     in f [] xs
 
-  let filter_presubs n xs =
+  let assoc_sub n xs name =
     let f n =
       let x = try assoc n !analyses_table with Not_found -> Printf.eprintf "filter_presubs: Analysis '%d' not registered.\n" n; failwith "filter_presubs" in
       let y = try assoc n xs with Not_found ->
@@ -169,7 +169,7 @@ struct
         Printf.eprintf "filter_presubs: Analysis '%s' (%d) not found.\n" x n; failwith "filter_presubs" in
       x, y
     in
-    map f (assoc n !dep_list)
+    assoc name (map f (assoc n !dep_list))
 
   let do_spawns ctx (xs:(varinfo * (int * lval option * exp list)) list) =
     let spawn_one v d =
@@ -243,8 +243,8 @@ struct
             ; edge   = ctx.edge
             ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
             ; emit   = (fun e -> emits := e :: !emits)
-            ; presub = filter_presubs n ctx.local
-            ; postsub= filter_presubs n post_all
+            ; presub = assoc_sub n ctx.local
+            ; postsub= assoc_sub n post_all
             ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
             ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
             ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -261,8 +261,8 @@ struct
             ; edge   = octx.edge
             ; ask    = (fun (type a) (q: a Queries.t) -> query octx q)
             ; emit   = (fun e -> emits := e :: !emits)
-            ; presub = filter_presubs n octx.local
-            ; postsub= filter_presubs n post_all
+            ; presub = assoc_sub n octx.local
+            ; postsub= assoc_sub n post_all
             ; global = (fun v      -> octx.global (v_of n v) |> g_to n |> obj)
             ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
             ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -300,8 +300,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -337,8 +337,8 @@ struct
           ; edge   = ctx.edge
           ; ask    = (fun (type b) (q: b Queries.t) -> query' asked' ctx q)
           ; emit   = (fun _ -> failwith "Cannot \"emit\" in query context.")
-          ; presub = filter_presubs n ctx.local
-          ; postsub= []
+          ; presub = assoc_sub n ctx.local
+          ; postsub= assoc_sub n []
           ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
           ; spawn  = (fun v d    -> failwith "Cannot \"spawn\" in query context.")
           ; split  = (fun d es   -> failwith "Cannot \"split\" in query context.")
@@ -384,8 +384,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun _ -> failwith "Cannot \"emit\" in access context.")
-        ; presub = filter_presubs n ctx.local
-        ; postsub= []
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n []
         ; global = (fun v         -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun v d       -> failwith "part_access::spawn")
         ; split  = (fun d es      -> failwith "part_access::split")
@@ -412,8 +412,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -446,8 +446,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -480,8 +480,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -515,8 +515,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -550,8 +550,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -585,8 +585,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -620,8 +620,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -655,8 +655,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -689,8 +689,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> splits := (n,(repr d,es)) :: !splits)
@@ -720,8 +720,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun _ -> failwith "Cannot \"emit\" in enter context.")
-        ; presub = filter_presubs n ctx.local
-        ; postsub= []
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n []
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun _ _    -> failwith "Cannot \"split\" in enter context." )
@@ -751,8 +751,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun l v a  -> spawns := (v,(n,l,a)) :: !spawns)
         ; split  = (fun d es   -> failwith "Cannot \"split\" in combine context.")
@@ -782,8 +782,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= []
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n []
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun v d    -> failwith "Cannot \"spawn\" in threadenter context.")
         ; split  = (fun d es   -> failwith "Cannot \"split\" in threadenter context.")
@@ -811,8 +811,8 @@ struct
         ; edge   = ctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query ctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n ctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n ctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> ctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun v d    -> failwith "Cannot \"spawn\" in threadspawn context.")
         ; split  = (fun d es   -> failwith "Cannot \"split\" in threadspawn context.")
@@ -829,8 +829,8 @@ struct
         ; edge   = fctx.edge
         ; ask    = (fun (type a) (q: a Queries.t) -> query fctx q)
         ; emit   = (fun e -> emits := e :: !emits)
-        ; presub = filter_presubs n fctx.local
-        ; postsub= filter_presubs n post_all
+        ; presub = assoc_sub n fctx.local
+        ; postsub= assoc_sub n post_all
         ; global = (fun v      -> fctx.global (v_of n v) |> g_to n |> obj)
         ; spawn  = (fun v d    -> failwith "Cannot \"spawn\" in threadspawn context.")
         ; split  = (fun d es   -> failwith "Cannot \"split\" in threadspawn context.")