### Changes
1. Change `presub` and `postsub` from eagerly constructed assoc lists to lazy functions. This avoids the expensive (~5% of total runtime) `filter_presubs`, which constructed those lists for each transfer function, but they aren't even used by anything other than ARINC and OSEK. This slowdown was my original motivation for getting rid of these (#460), but now that they aren't slowing down the usual analyses, they're not so much of a problem.
2. Reimplement `unop_fold` and `binop_fold` of `MCP`'s `Printable` and `Lattice` using `fold_left2` and `fold_left3` respectively, by folding the domain elements simultaneously with the list of activated analysis `Spec`s. This avoids any `assoc` in those operations. To avoid nasty segfaults, there are assertions to check that the component IDs still match up.
3. Reimplement `MCP`'s `combine` and `threadspawn` by folding over both domain elements (and optional contexts) simultaneously.
4. Replace `MCPRegistry` global assoc lists with `Hashtbl`s for faster lookup where still needed.

### TODO
- [x] Benchmark.
- [x] Figure out what Gobview needs the old `MCP.analyses_table` for, which was just a list of (analysis ID, analysis name) pairs.