@michael-schwarz 2022-01-28
>  Figure out what Gobview needs the old `MCP.analyses_table` for, which was just a list of (analysis ID, analysis name) pairs.

Can you shed some light on that @keremc?