@sim642 2022-01-28
I'm still running sv-benchmarks, but here's ypbind_comb.c analysis times.

#### Before
```
TOTAL                           8.226 s
  parse                           0.055 s
  convert to CIL                  0.028 s
  analysis                        8.144 s
    global_inits                    0.002 s
    solving                         7.453 s
      S.Dom.equal                     0.027 s
      postsolver                      1.815 s
        access compare                  0.055 s
    warn_global                     0.669 s
      access                          0.668 s
        group_may_race                  0.613 s
          bfs                             0.613 s
            access compare                  0.330 s
            may_race                        0.109 s
        may_race                        0.001 s
        access compare                  0.004 s
```

#### After
```
TOTAL                           7.699 s
  parse                           0.054 s
  convert to CIL                  0.026 s
  analysis                        7.620 s
    global_inits                    0.002 s
    solving                         7.074 s
      S.Dom.equal                     0.026 s
      postsolver                      1.774 s
        access compare                  0.056 s
    warn_global                     0.521 s
      access                          0.520 s
        group_may_race                  0.459 s
          bfs                             0.458 s
            access compare                  0.234 s
            may_race                        0.078 s
        may_race                        0.001 s
        access compare                  0.004 s
```