@sim642 2022-01-28
Here's the before vs after runtime on sv-benchmarks:
![image](https://user-images.githubusercontent.com/378740/151574042-e481672e-4ce0-47ca-bd81-40d919ec28f0.png)
[Full results table](https://goblint.cs.ut.ee/results/61-all-fast-mcp-no-assoc-after/results/results.2022-01-28_15-22-21.table.html#/scatter?toolY=1&columnY=1&regression=Linear&scaling=Linear&line=2&results=All&filter=0(0*status*(status(notIn(ERROR%20(verify),EXCEPTION%20(Stack%20overflow),OUT%20OF%20MEMORY,TIMEOUT,TIMEOUT%20(true))))),1(0*status*(status(notIn(ERROR%20(verify),EXCEPTION%20(Stack%20overflow),OUT%20OF%20MEMORY,TIMEOUT,TIMEOUT%20(true))))))

**That's a whopping ~19% improvement on average!**