@sim642 2022-01-31
I think I managed to make the Gobview adaptation. Function entry/return states and global states at least display correctly, statement nodes for some reason don't, but it's already that way on master: https://github.com/goblint/gobview/issues/6.