@keremc 2022-02-01
@michael-schwarz I don't know if this still holds, but I think the analysis ID used to be just an incremental ID and depended on the module initialization order. As a result, the ID was not stable across native and JS platforms, so I used `MCP.analyses_table` to reorder the analysis modules and reassign the IDs. At some point, the IDs ended up becoming stable across platforms somehow (maybe because of the `-linkall` flag), but I didn't remove the code for reordering.