@michael-schwarz 2022-02-01
Does https://github.com/goblint/gobview/compare/3acd89c9ec9b0714433ae3b9aa82d596a5c34aac...d0d5b065be6a763f78987e89172abc04afdc7dd5#diff-22c5afd13651a2ae673bd24ed0234d0b09f51d2ede9fd99236875dbc8715fd73 sufficienlty address this issue then? At least at first glance, it seems like it. Or would we need more code for that?